+++ Pinky/Pinky/PNKConcatenation.mm
@@ -220,8 +220,9 @@ secondaryInputFeatureChannels:(NSUInteger)secondaryInputFeatureChannels
   LTParameterAssert(primaryInputSize.width == secondaryInputSize.width &&
                     primaryInputSize.height == secondaryInputSize.height, @"Primary and secondary "
                     "inputs must have same width and height, got (%lu, %lu) and (%lu, %lu)",
-                    primaryInputSize.width, primaryInputSize.height, secondaryInputSize.width,
-                    secondaryInputSize.height);
+                    (unsigned long)primaryInputSize.width, (unsigned long)primaryInputSize.height,
+                    (unsigned long)secondaryInputSize.width,
+                    (unsigned long)secondaryInputSize.height);
   return {
     .width = primaryInputSize.width,
     .height = primaryInputSize.height,
